+++ lib/clichaumeil.lib.php
@@ -43,6 +43,11 @@ function clichaumeilAdminPrepareHead()
 	$head[$h][2] = 'settings';
 	$h++;
 
+	$head[$h][0] = dol_buildpath("/clichaumeil/admin/commissions.php", 1);
+	$head[$h][1] = $langs->trans("CliChaumeilCommissions");
+	$head[$h][2] = 'commissions';
+	$h++;
+
 	$head[$h][0] = dol_buildpath("/clichaumeil/admin/about.php", 1);
 	$head[$h][1] = $langs->trans("About");
 	$head[$h][2] = 'about';